requirements.txt
APScheduler==3.10.1
Django==4.2.2
Django==3.2.19
djangorestframework==3.14.0
Pillow==9.4.0
Pillow==9.5.0
selenium==4.10.0
